nely_web/requirements.txt
asgiref
Django
gunicorn
psycopg2
sqlparse
whitenoise
djangorestframework
drf-spectacular
djangorestframework-simplejwt
django-filter
dj_database_url
django-cors-headers
supabase
pillow
django-nested-admin
unidecode
resend